Update patch set 1
Patch Set 1:

writepid is perfect; however I don't think this will work as written.

if I understand this correctly, there's a race between writing the pid and executing the service. as a result, there is a possibility that the child process may have created its own thread before the writepid is finished, resulting in that worker thread not being included in the appropriate cpuset. this is bad.

can we make the child process write its own pid instead of the parent?

Patch-set: 1
Label: Code-Review=0