Update patch set 2
Patch Set 2:

Seems fine, but are you sure you want to encode this functionality in init.rc instead of in init?  Why not have keywords to specify the behavior that you want, and then have init take care of the details?  That way you can change the implementation later without having to find every device that uses it an update their init.rc.

Patch-set: 2