Update patch set 1
Patch Set 1:

The context for why the lock is required is in logd/LogStatistics.cpp (not touched in this CL) where the helpers originate. The issue is that calls to find the pidToName and pidToUid will create entries in the table if logs have not been submitted by the pid to prefill the content. pidToName, pidToUid and uidToName all read the existing table and thus could suffer issues iterating the list if it is changing.

Patch-set: 1