Update patch set 1
Patch Set 1: Code-Review-1

dear sir,
ran into your change by accident. don't want to suggest a differnet change because I'm not in the code...sort of speak. but,
if you set t->sfd to be -1 although the next line can potentially crash and not be "reflected" in the result. perhaps you should check adb_close() status, or something? I'm not fully in the code. hope this helps.

    int fd = t->sfd;
    if (fd != -1) {
        t->sfd = -1;
        adb_close(fd);
    }

Patch-set: 1