Update patch set 1
Patch Set 1:

> dear sir,
 > ran into your change by accident. don't want to suggest a differnet
 > change because I'm not in the code...sort of speak. but,
 > if you set t->sfd to be -1 although the next line can potentially
 > crash and not be "reflected" in the result. perhaps you should
 > check adb_close() status, or something? I'm not fully in the code.
 > hope this helps.

Thanks for the review. The code By Design sets t->sfd to -1 before calling adb_close() just in case the fd causes adb_close() to crash, in which case since t->sfd was just set, no one else in the future can try to close it again (causing problems repeatedly). So in the crash case, the code By Design leaks the fd which is preferable to leaving it around to cause more problems.

But really, this change does not have real provisions for the case of arbitrary lines of code crashing since that is a very difficult (impossible?) way to program this kind of code.

I checked the rest of the codebase and the return value of adb_close() is never checked, so this change follows that convention. adb_close() is really just an alias for unix/posix close() which is often not checked in many codebases.

Patch-set: 1