Update patch set 1
Patch Set 1: -Code-Review

I'm nothing less than disappointed to learn this is the state of the code base, if this is the case (of contributors) not taking the small efforts to do the simple checks, how can we rely in the future on this platform to drive cars? 
I'm in total agreement with you on this specific change after the effort you put but I can't believe we let this go on like that....should I change to plus instead? Or should we change the code to check the status, document the function to return the relevant error and all that? I can only suggest because I can't spend too much time on the, your call!

Patch-set: 1
Reviewer: Tzur Sayag <1075933@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0