Update patch set 1
Patch Set 1:

For the fast wake interval when on charger, it is not an issue because the Charger or USB driver will hold the wakelock to prevent suspend.

For any other uevents, it is the driver's responsibility to hold wakelock but not healthd/eventpoll. If it needs to be processed immediately, then the driver should hold the wakelock to make sure it will be processed in time.

Yes, many processes will wait on the uevent, but only healthd register it as EPOLLWAKEUP to hold wakelock. Is this a design on purpose?

Patch-set: 1
Reviewer: LianWei Wang <1017041@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0