Update patch set 1
Patch Set 1:

Did this test actually work for you on Windows?

When I type the commands manually into a Windows console window (with my Unicode change) everything works correctly, but when I run the python test (from my Windows box), it creates files on my device that are not correct.

My guess is that there is some difficulty with Python on Windows and Unicode. A quick web search suggests not calling encode('utf-8'), but when I remove that, I get some other errors -- it seems like it then wants every string to be unicode. So basically, I don't know the quick fix to this.

BTW, it is a real pain to use test cases that use Asian languages because to get it to display properly in console windows, you have to change the "system locale" (and reboot, argh) which then allows a proper font to be used in console windows (I mention this in a comment in sysdeps_win32.cpp).

When I was manually testing, I did test with Chinese characters, but more often I used a string that does not require system reconfiguration.  Go to http://stackoverflow.com/questions/2492077/output-unicode-strings-in-windows-console-app and search for "ambul" to see the string that I used.

Patch-set: 1