Update patch set 1
Patch Set 1:

(6 comments)

Example Windows output:

> adb connect asdf
unable to connect to asdf:5555: No such host is known. (11001)

> adb connect 192.168.1.2
unable to connect to 192.168.1.2:5555: A connection attempt failed because the connected party did not properly respond after a period of time, or established connection failed because connected host has failed to respond. (10060)

> adb connect 192.168.1.3
unable to connect to 192.168.3.251:5555: No connection could be made because the target machine actively refused it. (10061)

Patch-set: 1