Update patch set 3
Patch Set 3:

(2 comments)

The protobuffer generation is actually kind of tricky: All the import path must be relative to the root of the git repo otherwise it does not work.

I tried passing --proto_path explicitely in LOCAL_PROTOC_FLAGS but it ended up generating the protobuf differently (??).

Fixing the proto includes to mention system/core/metrics seemed to be the best solution.

Note: I made the static library export system/core/metrics/* so that we can still import *.pb.h with the same path.

Patch-set: 3