Update patch set 1
Patch Set 1:

I'd also be ok with changing these from warnings to just calling fatal(). Let me know if you think that would be better.

The big idea is just that if there is an error here, it should have good reporting. I don't really have a strong opinion on if it should be a warning or a hard error.

Note that if there are two adb server processes, in addition to both trying to grab port 5037, we now have both trying to open the same log file.

Patch-set: 1
Reviewer: Spencer Low <1065256@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0