Update patch set 1
Patch Set 1: Code-Review+2

Found cases where we triggered the supposedly 'superfluous' code. Merges and leading checking would delete the watermark as part of the cleanup. I actually recorded a 10-fold (if average is .5ms, a 5ms processing period) spike 1/30 tries without this; whereas none with the added 'superfluous' code. The average remained about the same with or without.

We have to consider the spike as being 'bad', so we once again approve of what we have done here.

Patch-set: 1
Label: Code-Review=+2