Update patch set 1
Patch Set 1:

This approximates the classes needed for Action and Service to wire into the Parser class on https://android-review.googlesource.com/#/c/162655/

It also cleans up the handling of raw pointers and by using shared_ptr and unique_ptr.

Still has some work to be done, but out for early review.

It boots successfully on Volantis.

Patch-set: 1
Label: Verified=0