Update patch set 1
Patch Set 1:

> How are these values set? Is that something you have change to the
 > build scripts for?

I believe those properties are not used on Android yet.

I haven't landed anything that sets those values but they should be set in the product directory (I believe there is a mechanism for this already: PRODUCT_PROPERTY_OVERRIDES). The metrics daemon discards the metrics if nothing is set so this change won't make metrics_daemon crash.

Patch-set: 1