Update patch set 12
Patch Set 12: Verified+1

(5 comments)

I wanted to break out the map lookup since the code to check the min and max args is identical in Action and Service.  Let me know if the KeywordMap class makes sense here or if that's overkill compared to the duplication of code.

I have a static function_map_ in Action, meant to decouple builtins.cpp from action.cpp without passing a pointer in every instance of Action.  I'm open to suggestions here of other ways to accomplish this.

This is getting close to removing the singletons for Parser, ActionManager, and ServiceManager, and having simply an Init singleton, which is an end goal.

This tests successfully on N6 and N9.

Patch-set: 12
Label: Verified=+1