Update patch set 3
Patch Set 3:

FYI, I don't think this works as expected on Windows because a socket connect() to a closed port on Windows takes 1 second (instead of the expected 0 sec). Thus, to try ADB_LOCAL_TRANSPORT_MAX emulator ports takes 64 seconds, instead of 1 sec as on Linux.

Practically speaking, this means that if the emulator connection goes down for some reason, it will take 32 seconds on average before the connection is reestablished. That is long enough that the user might think that it will never come back and they'll just kill the emulator and it'll be as if this change was never implemented.

Anecdote: years ago when I first learned Android app programming, I spent a lot of time debugging my app in an emulator (why? because the OS source code matched exactly what was running, unlike my Galaxy Nexus device at the time) and after N minutes, the connection would die. I had no idea what was going on and just had to live with it. It was an awful experience. (I don't know if this still occurs and I don't know what the root cause was.)

Here is info on why Windows takes 1 second for a failed socket connect:

https://jira.mongodb.org/browse/PYTHON-356
https://support.microsoft.com/en-us/kb/175523
https://mail.python.org/pipermail/python-dev/2009-January/085054.html (many msgs in thread)

Anyway, it sounds like Yabin has a http://b bug to fix this a better way, so I'm hoping that will close this issue for Windows. Thanks.

Patch-set: 3