Update patch set 2
Patch Set 2:

> Restored
 > 
 > Re-opening

The objective is to get rid of system(). It's true that init is designed to reap processes, so we're not really doing anything wrong by not wait()ing for it.

As for the implementation, I'd argue for using the wrappers even if we have to call Release() (to prevent the object SIGKILLing the process on destruction).

Patch-set: 2