Update patch set 2
Patch Set 2:

using fork+exec & not reaping is equiv to system and gets rid of it

as for reaping it, the point of this particular code path is to not get wedged even if dbus is wedged, so we explicitly don't want to wait on it.  in which case the only recourse is to reparent.

Patch-set: 2