Update patch set 1
Patch Set 1:

please see

logcat -v printable

This gives us gibberish without, yes, but a better result than truncated content. I can not _insert_ content into the buffer as it is sized, and I do not want to spend time in the logger reallocating and reformatting. That is the job of logcat. logcat already has a mechanism for this.

0x80 is _not_ a utf8 character, that is why I chose it, easy to spot.

Patch-set: 1