Update patch set 1
Patch Set 1:

but the only reason you have truncation is because you're incorrectly treating this as a C string when it's really an arbitrary byte sequence. if this code used a std::string or just carried around the actual length itself, it wouldn't truncate anything and we'd have lossless display of the exact bytes in logcat.

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0