Update patch set 2
Patch Set 2:

In practice the monotonic time 'calculation' is off by the accuracy of healthd. healthd is a background task, and the periodic synchronization can head to a few ms of inaccuracy. I hope this works for you. See the followup CLs where I can give you us of accuracy if you need it by switching the logging over to use CLOCK_MONOTONIC instead of CLOCK_REALTIME with the disadvantage of not being able to recover the realtime components.

Patch-set: 2