Update patch set 2
Patch Set 2:

(16 comments)

I made this change mainly because there is a bug report on linux that some fd is out of range of select(). Maybe it is a file handler leak, but I think we'd better get rid of the limit of select(). Using ADB_TRACE to enable LOG() is a good choice, I will do it in a separate CL.

Patch-set: 2
Label: Verified=0