Update patch set 2
Patch Set 2:

The problem with assertions in init is that when they fail, there is no easy way to debug them except for through the serial port, which is not always available.  If you want to make some errors in init fatal I suggest adding a little UI screen to show the error with a timeout and then exit and let the kernel panic and reboot.

Patch-set: 2