Update patch set 1
Patch Set 1:

> > isn't there some way we can avoid building useless 32-bit crap
 > > though?
 > 
 > Do you want to deprecate HOST_PREFER_32_BIT ?

i'd love to. unfortunately, aiui, you guys use it to test 32-bit art on the host. (ideally you'd do that by running the 32-bit x86 *target* binaries on the host -- which we do for bionic testing -- but that doesn't work if (a) you're deliberately trying to test with glibc rather than bionic or (b) you can't create /system on your host so that the executable can find /system/bin/linker and the necessary dynamic libraries when it starts.)

until the Windows world moves to 64-bit, though, i'll admit that i can't really escape having to support 32-bit and 64-bit for host tools. it's really only hurting build time and disk usage (and the number of duplicate error messages).

it seems especially sad that there's no way to ask for a 32-bit host art without having to have all the host tools built 32-bit. no one *actually* wants a 32-bit linux adb or fastboot.

Patch-set: 1