Update patch set 1
Patch Set 1:

Asus has a requirement for roughly 200MB of log retention. Others have remarked that they felt 16MB may be too small, but have been willing to accept that 16MB is eons above the 256KB size for the in-memory log buffers and would contact me if they felt like they needed more. It provided between 24 hours to 5 days of logging with that size on N6 and N9 respectively; but other development/debug targets could exhaust The on-disk buffer in a few hours. The 'ring buffer' is on disk and not in memory. I kept the (-n) number the same to keep the directory search manageable/scaled. I upped the (-r) size of the individual log file split from 64KB to 1024MB. I do not expect any scaling issues from 1024MB files except the initial startup search to find the last timestamp; the cost of doing business.

This is only delivered for userdebug and eng, and is off by default requiring an action to turn it on.

Patch-set: 1