Update patch set 1
Patch Set 1:

> David, this is the reason I'd prefer to have the adb server close
 > sockets before the adb client.
 > 
 > Siva, I noticed that Android Studio sometimes makes frequent adb
 > connections (i.e. once a second). For those, I think Android Studio
 > is already letting the adb server close its sockets before Android
 > Studio, so you probably don't have this problem.

Hmm, the only reason this could possibly happen is if the list of clients keeps changing (new processes start up and die on the device), and Studio attempts to connect to the JDWP port of each of them. If you ever see this, and you don't see traffic on the "adb jdwp" command, then could you please file a bug on Studio?

Patch-set: 1