Update patch set 1
Patch Set 1:

> Seems reasonable, it would be difficult to match ssh behavior
 > exactly and this is probably closest to what a user would expect.

Oh, I figured out what was going on. OpenSSH doesn't try to read from stdin until select returns that there's data available, so if you just wait and watch without typing anything, it never triggers the SIGTTIN. If you hit a character, ssh races with the process in the foreground, which results in the signal only sometimes happening.

I think the ssh behavior is basically a bug, so this behavior is probably strictly superior?

Patch-set: 1