Update patch set 5
Patch Set 5:

> except that seemed wrong because i don't have anything new that i
 > actually want to output.

LinePrinter.PrintOnNewLine("") seems ok to move to the new line.

The reality is that plain printf("\n") would probably also work, but it is probably not preferable because if you're using LinePrinter you probably want to stick with that instead of alternating between LinePrinter and printf (in the corner case where LinePrinter has buffered things up, or the case where a new line has already occurred). Again, I'm new to the code, so that's just a quick read...

 > yeah, i'm assuming i need to rewrite it all anyway at this point.
 > i'm hoping that it'll actually be _easier_ if SyncConnection is
 > more in control. then it can worry about whether to keep the last
 > line or not. hopefully we'll find out next week...

Yep! :)

Patch-set: 5