Update patch set 1
Patch Set 1:

touch in logd_write.c only affects initialization, or reinitialization (logd crash/restart, can not happen).
touch in fake_log_devices.c only affects host, which may be moot.

The real fix (tm) is in log_is_loggable.c, and in both cases they affect cache protection and have a guessing algorithm to sniff the cache to deal with contention. In that case we could consider a relatively tight nnn(?)-retry tight poll of trylock/sched_yield loop and fail over to the guessing algorithm rather than the 5 second timedlock. I will explore how we can determine we are nested or in a signal handler to give it more intellect.

Patch-set: 1