Update patch set 8
Patch Set 8:

> Just did a fresh sync and downloaded this CL, I get the same error
 > as enh with `mm` and `mma` builds in system/core, but the full `m`
 > build works. Maybe a missing dependency that the full build happens
 > to pick up elsewhere?

My theory is that maybe the CHECK() I have in utf8.cpp somehow causes something to be pulled in that is different than before. It will take me a while to repro & investigate this further because when I run mma from system/core, it fails because I don't have a full AOSP tree (I'm in the process of slowly downloading).

Patch-set: 8