Update patch set 1
Patch Set 1:

This is the start of an effort to get fastboot running over UDP, this patch is just pre-work to create a generic interface, and future patches will supply a UDP implementation. of the Transport interface.

LMK if there's anyone else I should add to fastboot reviews, I'm not sure who the main owners are here.

Thanks!

Tests so far:
  * Linux: getvar, flash, flashall
  * Windows: getvar, flash
  * Mac: None yet

Patch-set: 1