Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

Seems reasonable to me, just one nit but I would say don't bother with it unless you have to re-upload a new patch set anyway.

Would it be reasonably straightforward to add unit tests for ParseCompleteUTF8()? If so it's probably fine to omit, but it would add a little peace-of-mind if it won't require too much pain.

Patch-set: 1
Label: Code-Review=+1