Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

Resolve klogd issue where (note timestamp on logd-reinit):

12-02 14:06:16.847801     0     0 I init    : (Loading properties from /data/local.prop took 0.00s.)
12-02 14:06:17.061722   183   204 D Cryptfs : Just triggered post_fs_data
12-02 22:06:17.070056     0     0 I init    : Starting service 'logd-reinit'...
12-02 14:06:17.075480     0     0 I init    : Not bootcharting.

Disruption by the post_fs_data change which has a race with klogd, changes the timezone from GMT to local. Klogd rewrites all the logged messages to correct them, but not the inflight ones. This is not a problem if we select monotonic time, which is immune to timezone.

Patch-set: 1
Label: Code-Review=+1
Label: Verified=+1