Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)
 > 
 > I feel like we've had this conversation before, but I can't
 > remember the outcome. Is there a reason why brillo doesn't use the
 > AID_OEM_RESERVED_START range instead of adding it to the core
 > Android UIDs? IIUC, these UIDs won't be used on non-brillo
 > devices....

You're correct, we did have this conversation in the past. The distinction that I'm making is between core Brillo services (the services that enable Google-mediated Brillo value-added stuff, like metrics or Weave or A/B updates) and third-party services. I proposed and added the OEM range so that Brillo OEMs and developers had a reserved UID range to use for their services, without necessarily having to modify this file themselves.

I also proposed a way to manage that UID range using passwd/group files in the filesystem, but that work is temporarily blocked on concerns around interactions with the ramdisk.

Given that there's only 4 or 5 core Brillo services (full list in linked bug), I think it would be a good idea for Brillo to be able to use new UIDs for these services with a descriptive name.

Patch-set: 1