Update patch set 1
Patch Set 1:

> metricsd follows the same (incorrect in my mind) pattern of having
 > a binder object itself driving the binder service. I think the code
 > would be much cleaner if BnMetricsdImpl::Run() is moved outside of
 > BnMetricsdImpl and made a stand-alone function that takes
 > sp<BnMetricsdImpl> as parameter. There is no reason why Run must be
 > a member of the actual binder object...

I agree, the pattern is not good. We're actually changing it for https://android-review.googlesource.com/#/c/186627/7 and moving run outside of the binder object.
Todd, do you mind if I fold this into the other CL?

Note: for metricsd, we actually have two threads: one that receives the binder calls and is supposed to  be as light as possible and one that does all the IO work (backing up the metrics to disk, uploading, gathering metadata about the system, etc..)

Patch-set: 1
Reviewer: Bertrand Simonnet <1076133@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0