Update patch set 1
Patch Set 1:

If you added the -v usec format modifier, you would see the timestamp for the entry was probably '01-08 16:27:11.913236', and your request to print from '01-08 16:27:11.913' was like specifying '01-08 16:27:11.913000000' and thus _before_ the first entry. _If_ the first entry was actually '01-08 16:27:11.913000000', then it would behave as you expected.

This can not be reconciled even with -v usec, since the timestamps are recorded with nsec precision. _If_ I add a -v nsec format modifier, then maybe we can talk about this subtlety in grave detail. I am not sure there is a value to -v nsec as it takes 5usec to log an entry, so the nsec are like 'random numbers'. If you use the liblog interface (as the jni layer for "events" buffers does), then this will not be a problem, since the timestamp is reported in its full glory there.

If you wish to continue using an exec to logcat, I do not think you need -v usec for Device Owner logging, the extra time precision is meaningless.

So this is expected and designed behavior.

Patch-set: 1