Update patch set 2
Patch Set 2:

(1 comment)

> One thing that might work is to break things in the other
 > direction. If we lie, and tell adbd that the terminal is 1 column
 > narrower than it actually is, I think it might fix more things than
 > it breaks. Programs that print however many characters are in a
 > row, and rely on solely on text wrapping without using explicit
 > newlines will break, but that seems unlikely?

I tried -1 on the reported size and the problem still occurred (similar to the animated GIF that I originally posted). My thinking is that this sucks, but not to worry about it too much since it is "only" cosmetic and not functional like the other issues we've been fixing.

Patch-set: 2