Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

(1 comment)

If we find compatibility problems, I am willing to provide a partial revert of the -errno return value adjustment to the API. I would prefer to get it out to 'bake' the environment. If you feel strongly that this is a bad idea, my rubber arm is easy to twist, and I can pull this API change.

As an observer, if we do back it out, we will not have the ability to determine _why_ a call failed to ashmem since errno can not be relied upon.

Patch-set: 1
Label: Code-Review=+1
Label: Verified=+1