Update patch set 2
Patch Set 2:

> CleanupAlert:
 > 
 > logd actually starts at post-fs, the post-fs-data is a legacy to
 > start after persistent properties were set to configure logd
 > properly. We should have removed that line in post-fs-data when we
 > added logd-reinit to perform the task of telling logd to re-read
 > the properties.
 > 
 > Starting debuggerd at post-fs would required an investigation,
 > since debuggerd could report information in logs without the
 > ability to stuff a tombstone. The fact that the /data/ directory is
 > missing, then appearing later may cause it headaches?
 > 
 > filing two bugs to look after this technical debt.

my assumption was that it was futile trying to go earlier because either:

  * logd hasn't started, so we're not going to write anything to logcat, or

  * /data isn't mounted to we can't write tombstones.

aiui logd starting is the event that happens first, so debuggerd is useless earlier than that?

(yes, this should have been in the checkin comment.)

it's definitely useful to log to logcat even if we can't write a tombstone --- some evidence is better than no evidence.

Patch-set: 2