Update patch set 1
Patch Set 1:

I'm just saying if we don't need to worry about the ordering, we can remove the start ... lines from init.rc.  In theory this is preferable for debuggerd, since the 64bit variant may not be on every system and we'll print an error message for not finding the service if its not present.  But beyond that it's just bikeshedding.

Patch-set: 1