Update patch set 1
Patch Set 1:

For such releasing ownership code pattern, a more accurate method is to add attributes to hashmapPut's parameters so that analyzer knows that an object is not lost. But that is accurate only if hashmapPut "always" take ownership of its parameters.
As we can see current users of hashmapPut have already complicated conditions and comments about freeing exactly the non-released objects, I try to keep the information as accurate as possible for the analyzer and match the code (comments).

Patch-set: 1