Update patch set 1
Patch Set 1:

It is a security issue, being able to start the service on "user" builds would result in logcat data being stored on the eMMC. This is another layer of protection (beyond the long list already applied) to prevent it from ever happening. logcat data is considered a high value target only because it is sometimes the wild west when it comes to content.

I can not imagine the exploit, but its root would probably be the 'unknown bug' in init.

Patch-set: 1