Update patch set 4
Patch Set 4:

Scaled back to simply enabling the bounds sanitizer for now. It would be nice to enable object-size, but it's going to take more work than simply flipping on a switch now. It did work fine in Marshmallow and not that long ago in master, but it's brittle as it depends so heavily on compiler optimizations like _FORTIFY_SOURCE and violations of the C/C++ standards that are technically incompatible with it are common.

Patch-set: 4
Reviewer: Daniel Micay <1069238@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0