Update patch set 1
Patch Set 1:

There are several placed in logd that perform a ro.build.type check.

I do not know how ro.build.type and ro.debuggable are different from a security perspective; do you require that we change the "user" check in the logger?

I get nervous about compile time checks ;-) because of heisenberg causing a bug to show up in "user" that does appear to exist in "userdebug" traced to a remake of a binary affecting size or timing. Behavior notwithstanding ... ;-}

Patch-set: 1