Update patch set 2
Patch Set 2: Code-Review+1

Regarding a test case, how is it that the emulator can cause adb reboot to hang in such a way that it is now detected (with keepalives), but in such a way that we can't replicate it for a test? Isn't emulator.exe just using OS sockets?

Patch-set: 2
Reviewer: Spencer Low <1065256@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1