Update patch set 2
Patch Set 2: Presubmit-Ready+1

> (I don't really have context and only skimmed the thread.)
 > 
 > > Totally agree. If BoringSSL would provide a separate library with
 > > only basic crypto, I'd totally go for that. I'm not aware of such
 > a
 > > thing existing, and I don't think it's plausible to create it at
 > > this point.
 > 
 > Looking at openssl-stubs, most of that comes from OpenSSL's (really
 > awful) ASN.1 and X.509 implementations. It has always been the plan
 > to separate that out. I've gotten quite far in that now. (Note this
 > is recent work. I don't know if Android's BoringSSL is still
 > behind.) I am very much intending to unship all that junk from
 > Chrome eventually. One thought has indeed been have the
 > dependencies enforced by pulling those modules into their own
 > library.
 > 
 > Even without that work, static linkers and -ffunction-sections are
 > magic. If all you need is SHA-256, crypto/sha has already basically
 > no dependencies.

After discussions with enh@, drewry@, zeuthen@, the conclusion is to rely on mincrypt as an interim solution, with the goal of moving to the BoringSSL crypto primitives as soon as we get a BoringSSL variant targeted at embedded platforms and/or basic BoringSSL crypto in a separate library. I've contact the BoringSSL team to see what they think on the latter.

Anyone willing to give me a +2 now? ;)

Patch-set: 2
Label: Presubmit-Ready=0