Update patch set 1
Patch Set 1:

(7 comments)

Hard hitting and verbose review, the only _real_ important stuff is use getSizeTArg instead of strtoll and error out if -t (but not -T) and -m are used in combination and the request to add at least one gTest.

I would like to hold back commit on this CL until after a --regex CL is in the patch stack.

Tell me if you are interested in fixing the current long flags that have an underscore in them (we can document ones with -, and keep an internal duplicate with _ if you are worried about legacy usage). If not, I will take lead on that.

Patch-set: 1