Update patch set 1
Abandoned

Abandoning this. This cure is worse than the disease. I'll make BoringSSL use the underscores. Though I would ask that you all eventually get rid of this macro or at least define it at the end. (I'll be sending a patch to cURL for this. But I'm less confident sysdeps.h can be safely moved around.)

Patch-set: 1
Status: abandoned