Update patch set 1
Patch Set 1: Code-Review-2

> Note that once a library is added to this list the linker will
 > always prefer platform version to the one provided by an app.

Right. We can't submit this as-is. A large chunk of Codeplay's tests use this, and unfortunately will need to be rewritten to not use it. I just want them to have a way to make progress on other issues.

Patch-set: 1
Reviewer: Stephen Hines <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2