Update patch set 1
Patch Set 1:

My comment on https://googleplex-android-review.git.corp.google.com/#/c/902415/ wasn't just me being passive-aggressive, I'm genuinely not sure if we need this anymore.  I'm OK with keeping the test if we fix it.

I'll defer to the Wear team's judgment about whether they want to fix or trash the test.  I personally haven't seen anything on the phone or tablet side to convince me that this is still an issue, but Wear might be different.  (FWIW the underlying bug that started this was inside the kernel timekeeper, so it wasn't a question of faulty OEM code.)

Patch-set: 1
Reviewer: Greg Hackmann <1015340@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=0