Update patch set 1
Patch Set 1:

(1 comment)

Just a note why I do not care about the arbitrary 256 byte buffer I selected. In other places in the android ecosystem the result could be serious, we may wish to make this a 'better example' despite the fact that failure is regarded as an acceptable option.

Patch-set: 1