Update patch set 2
Patch Set 2:

(2 comments)

In this case, you may be able to use ERANGE to your advantage, since this is cosmetic code to replace user with a string if it is shorter than 5 characters; we use numbers if longer to keep logprint output 'neat'. A buffer of 128 (sizeof(passwd) + 64 + 6 + margin) would be acceptable with getpwuid_r()?

Patch-set: 2
Label: Presubmit-Verified=0