Update patch set 2
Patch Set 2:

> I am concerned we have partners using this as well, cleansing this
 > may require a 'generation' and it may be too soon to deprecate
 > the,.

not at all. we need to set an example, and moving Nexus devices off this is the best way to do that.

 > Also, I do not want to populate /system/etc/passwd with these
 > immutable ids, it give the impression that they can be changed, and
 > they can not.

this part i agree with. i don't see any particular advantage to moving everything to files, and i definitely worry about it encouraging OEMs to accidentally/deliberately break stuff.

 > Retaining them in a _private_ header file might be a necessary evil
 > to permit the host tool to be constructed.

there's no such thing as private in the platform. you yourself are regularly responsibly for adding #include <private/...> all over the codebase. if we mean to clean this up -- and i think we do -- we need to just do it.

Patch-set: 2